+++ src/corpus/CorpusDelete.vue
@@ -2,6 +2,7 @@
 import { useRouter } from "vue-router";
 import useCorpusIdParam from "./corpusIdParam.composable";
 import ActionButton from "@/components/ActionButton.vue";
+import RouteButton from "@/components/RouteButton.vue";
 import Section from "@/components/Section.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 import useCorpus from "./corpus.composable";
@@ -27,9 +28,9 @@ async function doDelete() {
           {{ $t("corpus.delete") }}
         </ActionButton>
 
-        <router-link :to="`/corpus/${corpusId}`">
-          <ActionButton>{{ $t("cancel") }}</ActionButton>
-        </router-link>
+        <RouteButton :to="`/corpus/${corpusId}`">{{
+          $t("cancel")
+        }}</RouteButton>
       </div>
     </Section>
   </PendingContent>