+++ src/corpus/config/CorpusConfiguration.vue
@@ -84,12 +84,10 @@
     </FormKit>
     <div class="flex justify-center">
       <PendingContent :on="`corpus/${corpusId}`">
-        <router-link :to="`/corpus/${corpusId}/delete`">
-          <ActionButton class="button-danger">
-            <icon :icon="['far', 'trash-can']" class="mr-1" />
-            {{ $t("corpus.delete") }}
-          </ActionButton>
-        </router-link>
+        <RouteButton :to="`/corpus/${corpusId}/delete`" class="button-danger">
+          <icon :icon="['far', 'trash-can']" class="mr-1" />
+          {{ $t("corpus.delete") }}
+        </RouteButton>
       </PendingContent>
     </div>
   </div>
@@ -100,7 +98,7 @@ import { computed } from "vue";
 import { useRouter } from "vue-router";
 import { useI18n } from "vue-i18n";
 import useCorpusIdParam from "@/corpus/corpusIdParam.composable";
-import ActionButton from "@/components/ActionButton.vue";
+import RouteButton from "@/components/RouteButton.vue";
 import Section from "@/components/Section.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 import useConfig from "./config.composable";